Add support for "KOFI" build type in PublishDataExtension
The PublishDataExtension.kt file was updated to accommodate a new "KOFI" build type. This change checks for an environmental variable "KOFI", and if set to "true", returns "KOFI" as the build type. This expands the variety of build types our project can handle, making it more flexible for different deployment environments.