Feature: project cleanup (#31)
* Add warning for non-stable repositories in PublishDataExtension

A warning message has been added in the addRepo function of the PublishDataExtension class. This serves as a caution whenever a non-stable repository is registered as the first one, encouraging users to register repositories in the correct order to maximize stable builds.

* Replace print statements with logger in PublishDataExtension

The print statements in the `PublishDataExtension` class for adding repository, tasks, and components have been changed to use the debug logging function to create more organized and detailed logs. In addition, for building on local branches and for non-public builds, warning logs have been used instead of print statements, improving clarity and error visibility.

* Add specific addRepo functions in PublishDataExtension

Introduced separate functions namely `addMainRepo`, `addMasterRepo`, `addDevRepo`, and `addSnapshotRepo` in the `PublishDataExtension` class for adding specific repository types. Notably, these changes simplify the code and eliminate duplication by replacing direct calls of `addRepo` function in `useGitlabReposForProject`, `useEldoNexusRepos`, and `useInternalEldoNexusRepos` with new specific functions.

* IllegalStateException is thrown if 'repos' is empty, ensuring that a repository is always defined.

* Update log messages in PublishDataExtension

Refactored the log messages in the PublishDataExtension file. Instead of using the standard println for logging, it's been changed to use the project's logger, which enhances readability and standardizes logging practice. This change includes the warning when the repository could not detect the release type and debug message when a release is detected.

* Add warning log for version changes in PublishDataExtension.kt

Altered the code in PublishDataExtension.kt to include a warning log when the project version gets cleaned/changed. The log advises consideration of removing any DEV or SNAPSHOT suffix. This change not only improves readability but also standardizes the logging procedure when versions are altered.

* Add debug log message for unspecified project version

In the PublishDataExtension.kt file, a debug log message has been added to notify when no version is set in the project and it defaults to the root project's version. This log is intended to provide clarity on version specifications and improve the oversight of the project's version handling.

* Add warning log for non-existent branch file

A warning log message has been added in PublishDataExtension.kt to indicate when a file doesn't exist due to undefined local branch. This warning helps maintain the traceability of commits, particularly when a branch is referred in the project but the corresponding file is missing.

* Add warning log for missing branch file

A warning log has been added within the determineLocalBranchInternal() function in the PublishDataExtension.kt file. This logs a warning when the .git/HEAD file doesn't exist, indicating an undefined local branch. This will help maintain traceability and identify references to non-existent branch files.

* Simplify conditions in getBuildType function

Revised the conditions in the getBuildType function within PublishDataExtension.kt file to be more concise. Removed redundant return statements and kept only the necessary environment variables checks which affect build types. This makes the code neater and easier to read.

* Bump version to 1.4.0

* Update Readme instruction for setup

The Readme file has been updated with simpler and clearer instructions for setting up and configuring repositories. A separate section explaining the configuration for each repository type has been created. This provides a step-by-step guide for general settings, adding repositories, Eldonexus, and GitLab configuration, making the setup process easier for users.

* Add support for "KOFI" build type in PublishDataExtension

The PublishDataExtension.kt file was updated to accommodate a new "KOFI" build type. This change checks for an environmental variable "KOFI", and if set to "true", returns "KOFI" as the build type. This expands the variety of build types our project can handle, making it more flexible for different deployment environments.

* Update Readme.md with build data and repository configurations

Enhanced the Readme.md with instructions on how to use publishData to generate a `build.data` file and updated repositories configurations. This provides users with more detail about the build process and methods to add additional values, thereby guiding them to properly configure their repositories and understand the build information.